requirements.txt
# Core dependencies
gradio>=4.0.0
boto3>=1.34.0
botocore>=1.34.0

# AI and ML libraries
langchain>=0.1.0
langchain-aws>=0.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
typing-extensions>=4.8.0

# Development and testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

python-docx>=1.1.0
fpdf2>=2.7.0